ci: clean up (#1)
